+++ plugins/Passbolt/DirectorySync/src/Actions/Traits/GroupUsersSyncTrait.php
@@ -308,6 +308,9 @@ trait GroupUsersSyncTrait
     {
         foreach ($userIdsToAdd as $userId) {
             $u = $this->Users->get($userId);
+            // For each round, we keep a track of previous groups users so that we can revert in case of error
+            // (and not keep the ones that generated errors in the list. This would create false positives for the next elements).
+            $initialGroupsUsers = $group->groups_users;
             try {
                 $newGroupUsers = $this->GroupsUsers->patchEntitiesWithChanges(
                     $group['groups_users'],
@@ -326,6 +329,9 @@ trait GroupUsersSyncTrait
                     Alias::STATUS_ERROR,
                     $error
                 ));
+
+                $group->groups_users = $initialGroupsUsers;
+                continue;
             } catch (\Exception $exception) {
                 $error = new SyncError($group, $exception);
                 $this->addReportItem(new ActionReport(
@@ -335,6 +341,9 @@ trait GroupUsersSyncTrait
                     Alias::STATUS_ERROR,
                     $error
                 ));
+
+                $group->groups_users = $initialGroupsUsers;
+                continue;
             }
 
             $saveResult = $this->Groups->save($group);
@@ -356,6 +365,12 @@ trait GroupUsersSyncTrait
                     Alias::STATUS_IGNORE,
                     $group
                 ));
+
+                $group->groups_users = $initialGroupsUsers;
+                continue;
+            }
+
+            if (empty($saveResult->groups_users)) {
                 continue;
             }
 
@@ -392,6 +407,7 @@ trait GroupUsersSyncTrait
                 continue;
             }
 
+            $initialGroupsUsers = $group->groups_users;
             try {
                 $newGroupUsers = $this->GroupsUsers->patchEntitiesWithChanges(
                     $group->groups_users,
@@ -410,6 +426,8 @@ trait GroupUsersSyncTrait
                     Alias::STATUS_IGNORE,
                     $error
                 ));
+
+                $group->groups_users = $initialGroupsUsers;
                 continue;
             }
 
@@ -424,6 +442,8 @@ trait GroupUsersSyncTrait
                 }
                 $error = new SyncError($group, new \Exception($msg));
                 $this->addReportItem(new ActionReport($msg, Alias::MODEL_GROUPS_USERS, Alias::ACTION_DELETE, Alias::STATUS_ERROR, $error));
+
+                $group->groups_users = $initialGroupsUsers;
                 continue;
             }